@@ -5,8 +5,9 @@
   "angular-cli": {},
   "scripts": {
     "ng": "ng",
+    "postinstall": "node patch.js",
     "build:aot": "ng build --prod -sm",
-    "start": "ng serve --aot --proxy-config proxy.conf.json --host 0.0.0.0",
+    "start": "ng serve --aot --proxy-config proxy.conf.json",
     "start.aot": "ng serve --aot --proxy-config proxy.conf.json --host ideal.com",
     "i18n": "ng xi18n --output-path src/locale --locale en",
     "lint": "tslint \"src/**/*.ts\" --project src/tsconfig.json --type-check && tslint \"e2e/**/*.ts\" --project e2e/tsconfig.json --type-check",
@@ -17,25 +18,23 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.6",
-    "@angular/cdk": "5.2.5",
-    "@angular/common": "^5.2.6",
-    "@angular/compiler": "^5.2.6",
-    "@angular/core": "^5.2.6",
-    "@angular/flex-layout": "v5.0.0-beta.14",
-    "@angular/forms": "^5.2.6",
-    "@angular/http": "^5.2.6",
-    "@angular/material": "5.2.5",
-    "@angular/platform-browser": "^5.2.6",
-    "@angular/platform-browser-dynamic": "^5.2.6",
-    "@angular/platform-server": "^5.2.6",
-    "@angular/router": "^5.2.6",
-    "@ngtools/webpack": "1.10.2",
-    "@nguniversal/express-engine": "^5.0.0-beta.5",
-    "@nguniversal/module-map-ngfactory-loader": "^5.0.0-beta.5",
+    "@angular/animations": "^6.0.3",
+    "@angular/cdk": "^6.2.0",
+    "@angular/common": "^6.0.3",
+    "@angular/compiler": "^6.0.3",
+    "@angular/core": "^6.0.3",
+    "@angular/flex-layout": "^6.0.0-beta.15",
+    "@angular/forms": "^6.0.3",
+    "@angular/http": "^6.0.3",
+    "@angular/material": "^6.2.0",
+    "@angular/platform-browser": "^6.0.3",
+    "@angular/platform-browser-dynamic": "^6.0.3",
+    "@angular/platform-server": "^6.0.3",
+    "@angular/router": "^6.0.3",
+    "@auth0/angular-jwt": "^2.0.0",
     "angular-froala-wysiwyg": "^2.7.2",
-    "angular2-jwt": "*",
-    "angular2-moment": "*",
+    "angular2-jwt": "^0.2.3",
+    "angular2-moment": "^1.9.0",
     "asn1js": "^2.0.12",
     "classlist.js": "*",
     "core-js": "^2.4.1",
@@ -45,7 +44,8 @@
     "froala-editor": "^2.7.5",
     "hammerjs": "*",
     "intl": "*",
-    "jszip": "^3.1.3",
+    "jszip": "^3.1.5",
+    "net": "^1.0.2",
     "ngx-infinite-scroll": "*",
     "node": "0.0.0",
     "node-forge": "^0.7.1",
@@ -53,8 +53,10 @@
     "nonce": "^1.0.4",
     "pbkdf2": "^3.0.9",
     "pkijs": "2.1.27",
-    "rxjs": "5.5.2",
-    "stompjs": "*",
+    "rxjs": "6.2.0",
+    "rxjs-compat": "^6.0.0-rc.0",
+    "stompjs": "^2.3.3",
+    "stream": "^0.0.2",
     "text-encoding": "^0.6.4",
     "ts-helpers": "^1.1.1",
     "ts-loader": "^3.5.0",
@@ -62,9 +64,10 @@
     "zone.js": "*"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.4",
-    "@angular/compiler-cli": "^5.2.6",
-    "@angular/language-service": "^5.2.6",
+    "@angular-devkit/build-angular": "~0.7.0-beta.1",
+    "@angular/cli": "^6.0.7",
+    "@angular/compiler-cli": "^6.0.3",
+    "@angular/language-service": "^6.0.3",
     "@types/asn1js": "0.0.1",
     "@types/crypto-js": "^3.1.33",
     "@types/file-saver": "0.0.1",
@@ -88,6 +91,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~3.0.4",
     "tslint": "~5.3.2",
-    "typescript": "2.6.2"
+    "typescript": "2.7.2"
   }
 }
