@@ -1,6 +1,7 @@
-import {ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot} from "@angular/router";
-import {Observable} from "rxjs/Observable";
-import {Injectable} from "@angular/core";
+import {ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot} from '@angular/router';
+import {Observable} from 'rxjs';
+import {Injectable} from '@angular/core';
+
 /**
  * Created by Viki on 5/13/2017.
  */
@@ -8,17 +9,17 @@ import {Injectable} from "@angular/core";
 @Injectable()
 export class AccessFromUrlNotAllowedGuard implements CanActivate {
 
-    constructor(private router: Router) {}
+  allow = false;
 
-    allow: boolean = false;
+  constructor(private router: Router) {
+  }
 
-    canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): boolean | Observable<boolean> | Promise<boolean> {
-        if (this.allow) {
-            this.allow = !this.allow;
-            return !this.allow;
-        }
-        this.router.navigate(["/home"]);
-        return this.allow;
+  canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): boolean | Observable<boolean> | Promise<boolean> {
+    if (this.allow) {
+      this.allow = !this.allow;
+      return !this.allow;
     }
-
+    this.router.navigate(['/home']);
+    return this.allow;
+  }
 }
