@@ -1,47 +1,47 @@
-import {Injectable} from "@angular/core";
-import {SecurityProfile} from "../../../domain/model/security/security-profile";
-import {JwtSecurityContext} from "../../authentication/jwt/jwt-security-context.service";
-import {KeysService} from "../keys/keys.service";
-import {CryptographicOperations} from "../cryptographic-operations/cryptographic-operations";
-import {UserService} from "../../../domain/services/user/user.service";
-import {CertificateService} from "../../../domain/services/certificate/certificate.service";
-import {HelperService} from "../helper.service";
-import {ParserPemService} from "../parsers/parser-pem.service";
-import {NoticeService} from "../../../domain/services/notice/notice.service";
-import {PaymentRequestPhaseData} from "../../../domain/model/security/data/payment-request-phase-data";
-import {ProtocolSession} from "../../../domain/model/security/protocol-session";
-import {ProtocolTransactionStepNotice} from "../../../domain/model/security/notices/protocol-transaction-step-notice";
-import {ProtocolTransactionStepNoticeConstructor} from "./protocol-transaction-step-notice-constructor.service";
-import {ProtocolTransactionStepThreeNotice} from "../../../domain/model/security/notices/protocol-transaction-step-three-notice";
-import {ProtocolTransactionStepOneNotice} from "../../../domain/model/security/notices/protocol-transaction-step-one-notice";
-import {ProtocolTransactionStepTwoNotice} from "../../../domain/model/security/notices/protocol-transaction-step-two-notice";
-import {ProtocolParticipantOneSessionData} from "../../../domain/model/security/protocol-participant-one-session-data";
-import {SimpleSecurityProfile} from "../../../domain/model/security/simple-security-profile";
-import {SecurityProfileConstructorService} from "./security-profile-constructor.service";
-import {SimpleCryptographicOperations} from "../cryptographic-operations/simple-cryptographic-operations";
-import {AlgorithmService} from "../algorithms/algorithms.service";
-import {SolutionService} from "../../../domain/services/solution/solution.service";
-import {Solution} from "../../../domain/model/ideas/solution";
-import {DecryptingService} from "../decrypting.service";
-import {EncryptingService} from "../encrypting.service";
-import {ProtocolTransactionService} from "../../../domain/services/protocol-transaction/protocol-transaction.service";
-import {ProtocolTransactionStepFourNotice} from "../../../domain/model/security/notices/protocol-transaction-step-four-notice";
-import {PaymentType} from "../../../domain/model/payment/payment_type";
-import {DigitalGoodsType} from "../../../domain/model/ideas/digital_goods_type";
-import {Epoid} from "../../../domain/model/security/data/epoid";
-import {ProtocolTransactionStepTwoDataRecipient} from "../../../domain/model/security/data/protocol-transaction-step-two-data-recipient";
-import {ProtocolTransactionStepOneDataRecipient} from "../../../domain/model/security/data/protocol-transaction-step-one-data-recipient";
-import {Observable} from "rxjs/Observable";
-import {PreviousNoticesData} from "../../../domain/model/security/data/previous-notices-data";
-import {CardInformation} from "../../../domain/model/security/data/card-information";
-import {ProtocolTransactionStepFourDataRecipient} from "../../../domain/model/security/data/protocol-transaction-step-four-data-recipient";
-import {ProtocolTransactionStepOneDataOriginator} from "../../../domain/model/security/data/protocol-transaction-step-one-data-originator";
-import {ProtocolTransactionStepFiveNotice} from "../../../domain/model/security/notices/protocol-transaction-step-five-notice";
-import {ProtocolTransactionStepFiveDataRecipient} from "../../../domain/model/security/data/protocol-transaction-step-five-data-recipient";
-import {UserCertificationService} from "../certificates/user-certification.service";
-import {ProtocolMessagesReconstructionService} from "./protocol-messages-reconstruction.service";
-import {CertificateOperationsService} from "../certificates/certificate-operations.service";
-import {MatSnackBar, MatSnackBarConfig} from "@angular/material";
+import {Injectable} from '@angular/core';
+import {SecurityProfile} from '../../../domain/model/security/security-profile';
+import {JwtSecurityContext} from '../../authentication/jwt/jwt-security-context.service';
+import {KeysService} from '../keys/keys.service';
+import {CryptographicOperations} from '../cryptographic-operations/cryptographic-operations';
+import {UserService} from '../../../domain/services/user/user.service';
+import {CertificateService} from '../../../domain/services/certificate/certificate.service';
+import {HelperService} from '../helper.service';
+import {ParserPemService} from '../parsers/parser-pem.service';
+import {NoticeService} from '../../../domain/services/notice/notice.service';
+import {PaymentRequestPhaseData} from '../../../domain/model/security/data/payment-request-phase-data';
+import {ProtocolSession} from '../../../domain/model/security/protocol-session';
+import {ProtocolTransactionStepNotice} from '../../../domain/model/security/notices/protocol-transaction-step-notice';
+import {ProtocolTransactionStepNoticeConstructor} from './protocol-transaction-step-notice-constructor.service';
+import {ProtocolTransactionStepThreeNotice} from '../../../domain/model/security/notices/protocol-transaction-step-three-notice';
+import {ProtocolTransactionStepOneNotice} from '../../../domain/model/security/notices/protocol-transaction-step-one-notice';
+import {ProtocolTransactionStepTwoNotice} from '../../../domain/model/security/notices/protocol-transaction-step-two-notice';
+import {ProtocolParticipantOneSessionData} from '../../../domain/model/security/protocol-participant-one-session-data';
+import {SimpleSecurityProfile} from '../../../domain/model/security/simple-security-profile';
+import {SecurityProfileConstructorService} from './security-profile-constructor.service';
+import {SimpleCryptographicOperations} from '../cryptographic-operations/simple-cryptographic-operations';
+import {AlgorithmService} from '../algorithms/algorithms.service';
+import {SolutionService} from '../../../domain/services/solution/solution.service';
+import {Solution} from '../../../domain/model/ideas/solution';
+import {DecryptingService} from '../decrypting.service';
+import {EncryptingService} from '../encrypting.service';
+import {ProtocolTransactionService} from '../../../domain/services/protocol-transaction/protocol-transaction.service';
+import {ProtocolTransactionStepFourNotice} from '../../../domain/model/security/notices/protocol-transaction-step-four-notice';
+import {PaymentType} from '../../../domain/model/payment/payment_type';
+import {DigitalGoodsType} from '../../../domain/model/ideas/digital_goods_type';
+import {Epoid} from '../../../domain/model/security/data/epoid';
+import {ProtocolTransactionStepTwoDataRecipient} from '../../../domain/model/security/data/protocol-transaction-step-two-data-recipient';
+import {ProtocolTransactionStepOneDataRecipient} from '../../../domain/model/security/data/protocol-transaction-step-one-data-recipient';
+import {Observable} from 'rxjs';
+import {PreviousNoticesData} from '../../../domain/model/security/data/previous-notices-data';
+import {CardInformation} from '../../../domain/model/security/data/card-information';
+import {ProtocolTransactionStepFourDataRecipient} from '../../../domain/model/security/data/protocol-transaction-step-four-data-recipient';
+import {ProtocolTransactionStepOneDataOriginator} from '../../../domain/model/security/data/protocol-transaction-step-one-data-originator';
+import {ProtocolTransactionStepFiveNotice} from '../../../domain/model/security/notices/protocol-transaction-step-five-notice';
+import {ProtocolTransactionStepFiveDataRecipient} from '../../../domain/model/security/data/protocol-transaction-step-five-data-recipient';
+import {UserCertificationService} from '../certificates/user-certification.service';
+import {ProtocolMessagesReconstructionService} from './protocol-messages-reconstruction.service';
+import {CertificateOperationsService} from '../certificates/certificate-operations.service';
+import {MatSnackBar, MatSnackBarConfig} from '@angular/material';
 
 /**
  * Created by Viki on 2/21/2017.
