@@ -1,11 +1,7 @@
-import {Injectable} from "@angular/core";
-import {SecurityProfile} from "../../domain/model/security/security-profile";
-import {Observable} from "rxjs/Observable";
-import {KeysService} from "./keys/keys.service";
-import {ParserPemService} from "./parsers/parser-pem.service";
-import {SimpleSecurityProfile} from "../../domain/model/security/simple-security-profile";
-import {ProtocolSession} from "../../domain/model/security/protocol-session";
-import {UserService} from "../../domain/services/user/user.service";
+import {Injectable} from '@angular/core';
+import {ProtocolSession} from '../../domain/model/security/protocol-session';
+import {UserService} from '../../domain/services/user/user.service';
+
 /**
  * Created by Viki on 2/10/2017.
  */
