@@ -1,13 +1,15 @@
+import {catchError} from 'rxjs/operators';
 /**
  * Created by AKuzmanoski on 08/01/2017.
  */
-import {Injectable} from "@angular/core";
-import {ActivatedRouteSnapshot, Resolve, RouterStateSnapshot} from "@angular/router";
-import {Announcement} from "../../../domain/model/sharing/announcement";
-import {Observable} from "rxjs";
-import {AnnouncementService} from "../../../domain/services/announcement/announcement.service";
-import {ErrorHandlingService} from "../../../core/error-handling/error-handling.service";
-import {Response} from "@angular/http";
+import {Injectable} from '@angular/core';
+import {ActivatedRouteSnapshot, Resolve, RouterStateSnapshot} from '@angular/router';
+import {Announcement} from '../../../domain/model/sharing/announcement';
+import {Observable} from 'rxjs';
+import {AnnouncementService} from '../../../domain/services/announcement/announcement.service';
+import {ErrorHandlingService} from '../../../core/error-handling/error-handling.service';
+import {Response} from '@angular/http';
+
 @Injectable()
 export class AnnouncementFeedListResolverService implements Resolve<Announcement[]> {
   constructor(private announcementServcie: AnnouncementService, private errorHandlingService: ErrorHandlingService) {
@@ -18,14 +20,14 @@ export class AnnouncementFeedListResolverService implements Resolve<Announcement
   resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<Announcement[]>
     | Promise<Announcement[]>
     | Announcement[] {
-    let type: string = route.data["type"];
-    let pageSize: number = route.data["pageSize"];
-    let query: string = route.queryParams["query"];
+    const type: string = route.data["type"];
+    const pageSize: number = route.data["pageSize"];
+    const query: string = route.queryParams["query"];
     return this.announcementServcie.getAnnouncementList({
       type: type,
       query: query,
       offset: "0",
       limit: pageSize.toString()
-    }).catch((error: Response) => this.errorHandlingService.handleError(error));
+    }).pipe(catchError((error: Response) => this.errorHandlingService.handleError(error)));
   }
 }
