@@ -1,9 +1,11 @@
-import {Injectable} from "@angular/core";
-import {Resolve, ActivatedRouteSnapshot, RouterStateSnapshot} from "@angular/router";
-import {Idea} from "../../domain/model/ideas/idea";
-import {Observable} from "rxjs";
-import {IdeaService} from "../../domain/services/idea/idea.service";
-import {ErrorHandlingService} from "../../core/error-handling/error-handling.service";
+import {catchError} from 'rxjs/operators';
+import {Injectable} from '@angular/core';
+import {ActivatedRouteSnapshot, Resolve, RouterStateSnapshot} from '@angular/router';
+import {Idea} from '../../domain/model/ideas';
+import {Observable} from 'rxjs';
+import {IdeaService} from '../../domain/services/idea/idea.service';
+import {ErrorHandlingService} from '../../core/error-handling/error-handling.service';
+
 /**
  * Created by AKuzmanoski on 05/01/2017.
  */
@@ -13,9 +15,9 @@ export class IdeaResolverService implements Resolve<Idea> {
 
   }
 
-  resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<Idea>|Promise<Idea>|Idea {
-    let ideaId: number = +route.params["id"];
+  resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<Idea> | Promise<Idea> | Idea {
+    const ideaId: number = +route.params['id'];
 
-    return this.ideaService.getIdea(ideaId).catch((error: any) => this.errorHandlingService.handleError(error));
+    return this.ideaService.getIdea(ideaId).pipe(catchError((error: any) => this.errorHandlingService.handleError(error)));
   }
 }
